Update README.md
Teaser++ is an excellent point cloud registration algorithm. However, currently the official version only supports deployment on the Linux platform. Here, we have completed the operation and deployment of Teaser++ code on the Windows platform based on VS2019 and PCL1.12. You only need to download the code, configure the PCL library, and then you can directly obtain the running results of Teaser++.